pyproject.toml
[project]
name = "llm-quiz-god-mode"
version = "0.1.0"
description = "LLM Quiz God Mode"
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
    "fastapi",
    "uvicorn",
    "python-dotenv",
    "playwright",
    "tenacity",
    "httpx",
    "pandas",
    "numpy",
    "matplotlib",
    "pypdf",
    "beautifulsoup4",
    "requests",
    "openpyxl",
    "email-validator"
]

[tool.uv]
dev-dependencies = []
